Alerting: remove dead ExtraConfiguration.GetAlertmanagerConfig/Validate (#129763)
