+++ pkg/services/ngalert/api/tooling/definitions/alertmanager_test.go
@@ -96,78 +96,6 @@ func Test_RawMessageMarshaling(t *testing.T) {
 	})
 }
 
-func TestExtraConfiguration_Validate(t *testing.T) {
-	testCases := []struct {
-		name          string
-		config        ExtraConfiguration
-		expectedError string
-	}{
-		{
-			name: "valid configuration",
-			config: ExtraConfiguration{
-				Identifier: "test-config",
-				AlertmanagerConfig: `route:
-  receiver: default
-receivers:
-  - name: default`,
-			},
-		},
-		{
-			name: "empty identifier",
-			config: ExtraConfiguration{
-				Identifier:         "",
-				AlertmanagerConfig: `route: {receiver: default}`,
-			},
-			expectedError: "identifier is required",
-		},
-		{
-			name: "invalid YAML alertmanager config",
-			config: ExtraConfiguration{
-				Identifier:         "test-config",
-				AlertmanagerConfig: `invalid: yaml: content: [`,
-			},
-			expectedError: "failed to parse alertmanager config",
-		},
-		{
-			name: "missing route in alertmanager config",
-			config: ExtraConfiguration{
-				Identifier: "test-config",
-				AlertmanagerConfig: `receivers:
-  - name: default`,
-			},
-			expectedError: "no routes provided",
-		},
-		{
-			name: "missing receivers in alertmanager config",
-			config: ExtraConfiguration{
-				Identifier: "test-config",
-				AlertmanagerConfig: `route:
-  receiver: default`,
-			},
-			expectedError: "undefined receiver",
-		},
-		{
-			name: "empty alertmanager config",
-			config: ExtraConfiguration{
-				Identifier:         "test-config",
-				AlertmanagerConfig: "",
-			},
-			expectedError: "failed to parse alertmanager config",
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			err := tc.config.Validate()
-			if tc.expectedError == "" {
-				require.NoError(t, err)
-			} else {
-				require.ErrorContains(t, err, tc.expectedError)
-			}
-		})
-	}
-}
-
 // Regression test: upstream Mimir/Cortex-compat Alertmanagers return an empty,
 // null, or missing `alertmanager_config` when no config has been saved. Before
 // the fix, UnmarshalYAML/UnmarshalJSON left amSimple nil, which made the
